Refactorizado el codigo

+++ src/components/TodoCard.jsx
@@ -5,9 +5,12 @@ const TodoCard = (props) => {
   const [complete, setComplete] = useState(props.complete)
   const [isDeleted, setDelete] = useState(false)
 
+  const API = 'https://todo-list-lnahuelfb.herokuapp.com/todos/'
+  // const API = 'http://localhost:3001/todos/'
+
   const handleDelete = async () => {
     try {
-      await fetch(`https://todo-list-lnahuelfb.herokuapp.com/todos/${props.id}`, {
+      await fetch(`${API}${props.id}`, {
         method: 'DELETE',
         headers: {
           'Content-Type': 'application/json'
@@ -22,7 +25,7 @@ const TodoCard = (props) => {
 
   const handleComplete = async () => {
     try {
-      await fetch(`https://todo-list-lnahuelfb.herokuapp.com/todos/${props.id}`, {
+      await fetch(`${API}${props.id}`, {
         method: 'PATCH',
         headers: {
           'Content-Type': 'application/json'
@@ -37,29 +40,24 @@ const TodoCard = (props) => {
     }
   }
 
-  return (
-    <>
-      {
-        isDeleted
-          ? null
-          : (
-            <div className='container'>
-              <h3>{props.task}</h3>
-              <p>Contenido: {props.description}</p>
-              <p>Important: {props.important.toString()}</p>
-              <p>Creado: {props.date}</p>
-              <p>Complete:
-                {
-                  complete ? ' ✅' : ' ❌'
-                }
-              </p>
-              <button className='card-button' onClick={() => handleComplete()}>Completado</button>
-              <button className='card-button' onClick={() => handleDelete()}>Eliminar</button>
-            </div>
-            )
-    }
-    </>
+  if (isDeleted) {
+    return null
+  }
 
+  return (
+    <div className='container'>
+      <h3>{props.task}</h3>
+      <p>Contenido: {props.description}</p>
+      <p>Important: {props.important.toString()}</p>
+      <p>Creado: {props.date}</p>
+      <p>Complete:
+        {
+          complete ? ' ✅' : ' ❌'
+        }
+      </p>
+      <button className='card-button' onClick={() => handleComplete()}>Completado</button>
+      <button className='card-button' onClick={() => handleDelete()}>Eliminar</button>
+    </div>
   )
 }
